+++ HISTORY.md
@@ -78,6 +78,9 @@
     * German
     * French
 
+### 2019-05-20
+* Objects, Blocks added plus- and minus- keys to key pressed predicate and hat block 
+
 ### 2019-05-16
 * Objects: more refactoring to generalize projection extensions